+++ clang/lib/Driver/ToolChains/MSVC.cpp
@@ -964,6 +964,7 @@ static void TranslatePermissiveMinus(Arg *A, llvm::opt::DerivedArgList &DAL,
                                      const OptTable &Opts) {
   DAL.AddFlagArg(A, Opts.getOption(options::OPT__SLASH_Zc_twoPhase));
   DAL.AddFlagArg(A, Opts.getOption(options::OPT_foperator_names));
+  DAL.AddFlagArg(A, Opts.getOption(options::OPT_fno_ms_reference_binding));
 }
 
 llvm::opt::DerivedArgList *